Port SenderKeyStore to async bridge_fn
