+++ rust/bridge/ffi/src/lib.rs
@@ -666,190 +666,6 @@ pub unsafe extern "C" fn signal_decrypt_pre_key_message(
     })
 }
 
-type LoadSenderKey = extern "C" fn(
-    store_ctx: *mut c_void,
-    *mut *mut SenderKeyRecord,
-    *const SenderKeyName,
-    ctx: *mut c_void,
-) -> c_int;
-type StoreSenderKey = extern "C" fn(
-    store_ctx: *mut c_void,
-    *const SenderKeyName,
-    *const SenderKeyRecord,
-    ctx: *mut c_void,
-) -> c_int;
-
-#[repr(C)]
-#[derive(Copy, Clone)]
-pub struct FfiSenderKeyStoreStruct {
-    ctx: *mut c_void,
-    load_sender_key: LoadSenderKey,
-    store_sender_key: StoreSenderKey,
-}
-
-pub struct FfiSenderKeyStore {
-    store: FfiSenderKeyStoreStruct,
-}
-
-impl FfiSenderKeyStore {
-    fn new(store: *const FfiSenderKeyStoreStruct) -> Result<Self, SignalFfiError> {
-        Ok(Self {
-            store: *unsafe { store.as_ref() }.ok_or(SignalFfiError::NullPointer)?,
-        })
-    }
-}
-
-#[async_trait(?Send)]
-impl SenderKeyStore for FfiSenderKeyStore {
-    async fn store_sender_key(
-        &mut self,
-        sender_key_name: &SenderKeyName,
-        record: &SenderKeyRecord,
-        ctx: Context,
-    ) -> Result<(), SignalProtocolError> {
-        let ctx = ctx.unwrap_or(std::ptr::null_mut());
-        let result =
-            (self.store.store_sender_key)(self.store.ctx, &*sender_key_name, &*record, ctx);
-
-        if let Some(error) = CallbackError::check(result) {
-            return Err(SignalProtocolError::ApplicationCallbackError(
-                "store_sender_key",
-                Box::new(error),
-            ));
-        }
-
-        Ok(())
-    }
-
-    async fn load_sender_key(
-        &mut self,
-        sender_key_name: &SenderKeyName,
-        ctx: Context,
-    ) -> Result<Option<SenderKeyRecord>, SignalProtocolError> {
-        let ctx = ctx.unwrap_or(std::ptr::null_mut());
-        let mut record = std::ptr::null_mut();
-        let result =
-            (self.store.load_sender_key)(self.store.ctx, &mut record, &*sender_key_name, ctx);
-
-        if let Some(error) = CallbackError::check(result) {
-            return Err(SignalProtocolError::ApplicationCallbackError(
-                "load_sender_key",
-                Box::new(error),
-            ));
-        }
-
-        if record.is_null() {
-            return Ok(None);
-        }
-
-        let record = unsafe { Box::from_raw(record) };
-
-        Ok(Some(*record))
-    }
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn signal_create_sender_key_distribution_message(
-    obj: *mut *mut SenderKeyDistributionMessage,
-    sender_key_name: *const SenderKeyName,
-    store: *const FfiSenderKeyStoreStruct,
-    ctx: *mut c_void,
-) -> *mut SignalFfiError {
-    run_ffi_safe(|| {
-        if sender_key_name.is_null() || store.is_null() {
-            return Err(SignalFfiError::NullPointer);
-        }
-        let sender_key_name = native_handle_cast::<SenderKeyName>(sender_key_name)?;
-
-        let mut sender_key_store = FfiSenderKeyStore::new(store)?;
-        let mut csprng = rand::rngs::OsRng;
-
-        let skdm = expect_ready(create_sender_key_distribution_message(
-            &sender_key_name,
-            &mut sender_key_store,
-            &mut csprng,
-            Some(ctx),
-        ));
-
-        box_object::<SenderKeyDistributionMessage>(obj, skdm)
-    })
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn signal_process_sender_key_distribution_message(
-    sender_key_name: *const SenderKeyName,
-    sender_key_distribution_message: *const SenderKeyDistributionMessage,
-    store: *const FfiSenderKeyStoreStruct,
-    ctx: *mut c_void,
-) -> *mut SignalFfiError {
-    run_ffi_safe(|| {
-        let sender_key_name = native_handle_cast::<SenderKeyName>(sender_key_name)?;
-        let sender_key_distribution_message =
-            native_handle_cast::<SenderKeyDistributionMessage>(sender_key_distribution_message)?;
-        let mut sender_key_store = FfiSenderKeyStore::new(store)?;
-
-        expect_ready(process_sender_key_distribution_message(
-            sender_key_name,
-            sender_key_distribution_message,
-            &mut sender_key_store,
-            Some(ctx),
-        ))?;
-
-        Ok(())
-    })
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn signal_group_encrypt_message(
-    out: *mut *const c_uchar,
-    out_len: *mut size_t,
-    sender_key_name: *const SenderKeyName,
-    message: *const c_uchar,
-    message_len: size_t,
-    store: *const FfiSenderKeyStoreStruct,
-    ctx: *mut c_void,
-) -> *mut SignalFfiError {
-    run_ffi_safe(|| {
-        let sender_key_name = native_handle_cast::<SenderKeyName>(sender_key_name)?;
-        let message = as_slice(message, message_len)?;
-        let mut sender_key_store = FfiSenderKeyStore::new(store)?;
-        let mut rng = rand::rngs::OsRng;
-        let ctext = expect_ready(group_encrypt(
-            &mut sender_key_store,
-            &sender_key_name,
-            &message,
-            &mut rng,
-            Some(ctx),
-        ))?;
-        write_bytearray_to(out, out_len, ctext)
-    })
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn signal_group_decrypt_message(
-    out: *mut *const c_uchar,
-    out_len: *mut size_t,
-    sender_key_name: *const SenderKeyName,
-    message: *const c_uchar,
-    message_len: size_t,
-    store: *const FfiSenderKeyStoreStruct,
-    ctx: *mut c_void,
-) -> *mut SignalFfiError {
-    run_ffi_safe(|| {
-        let sender_key_name = native_handle_cast::<SenderKeyName>(sender_key_name)?;
-        let message = as_slice(message, message_len)?;
-        let mut sender_key_store = FfiSenderKeyStore::new(store)?;
-
-        let ptext = expect_ready(group_decrypt(
-            &message,
-            &mut sender_key_store,
-            &sender_key_name,
-            Some(ctx),
-        ))?;
-        write_bytearray_to(out, out_len, ptext)
-    })
-}
-
 #[no_mangle]
 pub unsafe extern "C" fn signal_sealed_session_cipher_encrypt(
     out: *mut *const c_uchar,
