+++ rust/bridge/jni/src/lib.rs
@@ -23,7 +23,6 @@ type JavaSessionStore<'a> = JObject<'a>;
 type JavaIdentityKeyStore<'a> = JObject<'a>;
 type JavaPreKeyStore<'a> = JObject<'a>;
 type JavaSignedPreKeyStore<'a> = JObject<'a>;
-type JavaSenderKeyStore<'a> = JObject<'a>;
 
 type JavaCiphertextMessage = jobject;
 
@@ -45,29 +44,6 @@ pub unsafe extern "C" fn Java_org_signal_client_internal_Native_IdentityKeyPair_
         env.set_long_array_region(result, 0, &tuple)?;
         Ok(result)
     })
-}
-
-/* SenderKeyName */
-
-fn sender_key_name_to_jobject<'a>(
-    env: &'a JNIEnv,
-    sender_key_name: &SenderKeyName,
-) -> Result<JObject<'a>, SignalJniError> {
-    let sender_key_name_class =
-        env.find_class("org/whispersystems/libsignal/groups/SenderKeyName")?;
-    let sender_key_name_ctor_args = [
-        JObject::from(env.new_string(sender_key_name.group_id()?)?).into(),
-        JObject::from(env.new_string(sender_key_name.sender_name()?)?).into(),
-        JValue::from(jint_from_u32(sender_key_name.sender_device_id())?),
-    ];
-
-    let sender_key_name_ctor_sig = "(Ljava/lang/String;Ljava/lang/String;I)V";
-    let sender_key_name_jobject = env.new_object(
-        sender_key_name_class,
-        sender_key_name_ctor_sig,
-        &sender_key_name_ctor_args,
-    )?;
-    Ok(sender_key_name_jobject)
 }
 
 fn protocol_address_to_jobject<'a>(
@@ -671,188 +647,6 @@ pub unsafe extern "C" fn Java_org_signal_client_internal_Native_SessionCipher_1D
             &mut prekey_store,
             &mut signed_prekey_store,
             &mut csprng,
-            None,
-        ))?;
-
-        to_jbytearray(&env, Ok(ptext))
-    })
-}
-
-pub struct JniSenderKeyStore<'a> {
-    env: &'a JNIEnv<'a>,
-    store: JObject<'a>,
-}
-
-impl<'a> JniSenderKeyStore<'a> {
-    fn new(env: &'a JNIEnv, store: JObject<'a>) -> Result<Self, SignalJniError> {
-        check_jobject_type(
-            &env,
-            store,
-            "org/whispersystems/libsignal/groups/state/SenderKeyStore",
-        )?;
-        Ok(Self { env, store })
-    }
-}
-
-impl<'a> JniSenderKeyStore<'a> {
-    fn do_store_sender_key(
-        &mut self,
-        sender_key_name: &SenderKeyName,
-        record: &SenderKeyRecord,
-    ) -> Result<(), SignalJniError> {
-        let sender_key_name_jobject = sender_key_name_to_jobject(self.env, sender_key_name)?;
-        let sender_key_record_jobject = jobject_from_native_handle(
-            self.env,
-            "org/whispersystems/libsignal/groups/state/SenderKeyRecord",
-            box_object::<SenderKeyRecord>(Ok(record.clone()))?,
-        )?;
-
-        let callback_args = [
-            sender_key_name_jobject.into(),
-            sender_key_record_jobject.into(),
-        ];
-        let callback_sig = "(Lorg/whispersystems/libsignal/groups/SenderKeyName;Lorg/whispersystems/libsignal/groups/state/SenderKeyRecord;)V";
-        call_method_checked(
-            self.env,
-            self.store,
-            "storeSenderKey",
-            callback_sig,
-            &callback_args[..],
-        )?;
-
-        Ok(())
-    }
-
-    fn do_load_sender_key(
-        &mut self,
-        sender_key_name: &SenderKeyName,
-    ) -> Result<Option<SenderKeyRecord>, SignalJniError> {
-        let sender_key_name_jobject = sender_key_name_to_jobject(self.env, sender_key_name)?;
-        let callback_args = [sender_key_name_jobject.into()];
-        let callback_sig = "(Lorg/whispersystems/libsignal/groups/SenderKeyName;)Lorg/whispersystems/libsignal/groups/state/SenderKeyRecord;";
-
-        let skr = get_object_with_native_handle::<SenderKeyRecord>(
-            self.env,
-            self.store,
-            &callback_args,
-            callback_sig,
-            "loadSenderKey",
-        )?;
-
-        Ok(skr)
-    }
-}
-
-#[async_trait(?Send)]
-impl<'a> SenderKeyStore for JniSenderKeyStore<'a> {
-    async fn store_sender_key(
-        &mut self,
-        sender_key_name: &SenderKeyName,
-        record: &SenderKeyRecord,
-        _ctx: Context,
-    ) -> Result<(), SignalProtocolError> {
-        Ok(self.do_store_sender_key(sender_key_name, record)?)
-    }
-
-    async fn load_sender_key(
-        &mut self,
-        sender_key_name: &SenderKeyName,
-        _ctx: Context,
-    ) -> Result<Option<SenderKeyRecord>, SignalProtocolError> {
-        Ok(self.do_load_sender_key(sender_key_name)?)
-    }
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn Java_org_signal_client_internal_Native_GroupSessionBuilder_1CreateSenderKeyDistributionMessage(
-    env: JNIEnv,
-    _class: JClass,
-    sender_key_name: ObjectHandle,
-    store: JavaSenderKeyStore,
-) -> ObjectHandle {
-    run_ffi_safe(&env, || {
-        let sender_key_name = native_handle_cast::<SenderKeyName>(sender_key_name)?;
-        let mut sender_key_store = JniSenderKeyStore::new(&env, store)?;
-        let mut csprng = rand::rngs::OsRng;
-
-        let skdm = expect_ready(create_sender_key_distribution_message(
-            &sender_key_name,
-            &mut sender_key_store,
-            &mut csprng,
-            None,
-        ))?;
-        box_object::<SenderKeyDistributionMessage>(Ok(skdm))
-    })
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn Java_org_signal_client_internal_Native_GroupSessionBuilder_1ProcessSenderKeyDistributionMessage(
-    env: JNIEnv,
-    _class: JClass,
-    sender_key_name: ObjectHandle,
-    sender_key_distribution_message: ObjectHandle,
-    store: JavaSenderKeyStore,
-) {
-    run_ffi_safe(&env, || {
-        let sender_key_name = native_handle_cast::<SenderKeyName>(sender_key_name)?;
-        let sender_key_distribution_message =
-            native_handle_cast::<SenderKeyDistributionMessage>(sender_key_distribution_message)?;
-        let mut sender_key_store = JniSenderKeyStore::new(&env, store)?;
-
-        expect_ready(process_sender_key_distribution_message(
-            sender_key_name,
-            sender_key_distribution_message,
-            &mut sender_key_store,
-            None,
-        ))?;
-        Ok(())
-    })
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn Java_org_signal_client_internal_Native_GroupCipher_1EncryptMessage(
-    env: JNIEnv,
-    _class: JClass,
-    sender_key_name: ObjectHandle,
-    message: jbyteArray,
-    store: JavaSenderKeyStore,
-) -> jbyteArray {
-    run_ffi_safe(&env, || {
-        let sender_key_name = native_handle_cast::<SenderKeyName>(sender_key_name)?;
-        let message = env.convert_byte_array(message)?;
-        let mut sender_key_store = JniSenderKeyStore::new(&env, store)?;
-
-        let mut rng = rand::rngs::OsRng;
-
-        let ctext = expect_ready(group_encrypt(
-            &mut sender_key_store,
-            &sender_key_name,
-            &message,
-            &mut rng,
-            None,
-        ))?;
-
-        to_jbytearray(&env, Ok(ctext))
-    })
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn Java_org_signal_client_internal_Native_GroupCipher_1DecryptMessage(
-    env: JNIEnv,
-    _class: JClass,
-    sender_key_name: ObjectHandle,
-    message: jbyteArray,
-    store: JavaSenderKeyStore,
-) -> jbyteArray {
-    run_ffi_safe(&env, || {
-        let sender_key_name = native_handle_cast::<SenderKeyName>(sender_key_name)?;
-        let message = env.convert_byte_array(message)?;
-        let mut sender_key_store = JniSenderKeyStore::new(&env, store)?;
-
-        let ptext = expect_ready(group_decrypt(
-            &message,
-            &mut sender_key_store,
-            &sender_key_name,
             None,
         ))?;
