+++ rust/bridge/jni/src/util.rs
@@ -26,63 +26,6 @@ pub fn jint_from_u32(value: Result<u32, SignalProtocolError>) -> Result<jint, Si
     }
 }
 
-pub fn check_jobject_type(
-    env: &JNIEnv,
-    obj: JObject,
-    class_name: &'static str,
-) -> Result<(), SignalJniError> {
-    if obj.is_null() {
-        return Err(SignalJniError::NullHandle);
-    }
-
-    let class = env.find_class(class_name)?;
-
-    if !env.is_instance_of(obj, class)? {
-        return Err(SignalJniError::BadJniParameter(class_name));
-    }
-
-    Ok(())
-}
-
-pub fn get_object_with_native_handle<T: 'static + Clone>(
-    env: &JNIEnv,
-    store_obj: JObject,
-    callback_args: &[JValue],
-    callback_sig: &'static str,
-    callback_fn: &'static str,
-) -> Result<Option<T>, SignalJniError> {
-    let rvalue = call_method_checked(env, store_obj, callback_fn, callback_sig, &callback_args)?;
-
-    let obj = match rvalue {
-        JValue::Object(o) => *o,
-        _ => {
-            return Err(SignalJniError::UnexpectedJniResultType(
-                callback_fn,
-                rvalue.type_name(),
-            ))
-        }
-    };
-
-    if obj.is_null() {
-        return Ok(None);
-    }
-
-    let handle = call_method_checked(env, obj, "nativeHandle", "()J", &[])?;
-    match handle {
-        JValue::Long(handle) => {
-            if handle == 0 {
-                return Ok(None);
-            }
-            let object = unsafe { native_handle_cast::<T>(handle)? };
-            Ok(Some(object.clone()))
-        }
-        _ => Err(SignalJniError::UnexpectedJniResultType(
-            "nativeHandle",
-            handle.type_name(),
-        )),
-    }
-}
-
 pub fn get_object_with_serialization(
     env: &JNIEnv,
     store_obj: JObject,
@@ -125,16 +68,5 @@ pub fn jobject_from_serialized<'a>(
     let class_type = env.find_class(class_name)?;
     let ctor_sig = "([B)V";
     let ctor_args = [JValue::from(to_jbytearray(env, Ok(serialized))?)];
-    Ok(env.new_object(class_type, ctor_sig, &ctor_args)?)
-}
-
-pub fn jobject_from_native_handle<'a>(
-    env: &'a JNIEnv,
-    class_name: &str,
-    boxed_handle: ObjectHandle,
-) -> Result<JObject<'a>, SignalJniError> {
-    let class_type = env.find_class(class_name)?;
-    let ctor_sig = "(J)V";
-    let ctor_args = [JValue::from(boxed_handle)];
     Ok(env.new_object(class_type, ctor_sig, &ctor_args)?)
 }
