+++ rust/bridge/shared/Cargo.toml
@@ -20,6 +20,7 @@ paste = "1.0"
 rand = "0.7.3"
 static_assertions = "1.1"
 scopeguard = "1.0"
+async-trait = "0.1.41"
 
 libc = { version = "0.2", optional = true }
 jni = { version = "0.17", optional = true }
