+++ rust/bridge/shared/src/ffi/mod.rs
@@ -14,6 +14,9 @@ pub(crate) use convert::*;
 mod error;
 pub use error::*;
 
+mod storage;
+pub use storage::*;
+
 pub use crate::support::expect_ready;
 
 pub fn run_ffi_safe<F: FnOnce() -> Result<(), SignalFfiError> + std::panic::UnwindSafe>(
