+++ rust/bridge/shared/src/jni/mod.rs
@@ -22,6 +22,9 @@ pub(crate) use convert::*;
 mod error;
 pub use error::*;
 
+mod storage;
+pub use storage::*;
+
 pub use crate::support::expect_ready;
 
 pub type ObjectHandle = jlong;
@@ -305,6 +308,74 @@ pub fn call_method_checked<'a>(
             Box::new(ThrownException::new(env, throwable)?),
         )
         .into())
+    }
+}
+
+pub fn jobject_from_native_handle<'a>(
+    env: &'a JNIEnv,
+    class_name: &str,
+    boxed_handle: ObjectHandle,
+) -> Result<JObject<'a>, SignalJniError> {
+    let class_type = env.find_class(class_name)?;
+    let ctor_sig = "(J)V";
+    let ctor_args = [JValue::from(boxed_handle)];
+    Ok(env.new_object(class_type, ctor_sig, &ctor_args)?)
+}
+
+pub fn check_jobject_type(
+    env: &JNIEnv,
+    obj: JObject,
+    class_name: &'static str,
+) -> Result<(), SignalJniError> {
+    if obj.is_null() {
+        return Err(SignalJniError::NullHandle);
+    }
+
+    let class = env.find_class(class_name)?;
+
+    if !env.is_instance_of(obj, class)? {
+        return Err(SignalJniError::BadJniParameter(class_name));
+    }
+
+    Ok(())
+}
+
+pub fn get_object_with_native_handle<T: 'static + Clone>(
+    env: &JNIEnv,
+    store_obj: JObject,
+    callback_args: &[JValue],
+    callback_sig: &'static str,
+    callback_fn: &'static str,
+) -> Result<Option<T>, SignalJniError> {
+    let rvalue = call_method_checked(env, store_obj, callback_fn, callback_sig, &callback_args)?;
+
+    let obj = match rvalue {
+        JValue::Object(o) => *o,
+        _ => {
+            return Err(SignalJniError::UnexpectedJniResultType(
+                callback_fn,
+                rvalue.type_name(),
+            ))
+        }
+    };
+
+    if obj.is_null() {
+        return Ok(None);
+    }
+
+    let handle = call_method_checked(env, obj, "nativeHandle", "()J", &[])?;
+    match handle {
+        JValue::Long(handle) => {
+            if handle == 0 {
+                return Ok(None);
+            }
+            let object = unsafe { native_handle_cast::<T>(handle)? };
+            Ok(Some(object.clone()))
+        }
+        _ => Err(SignalJniError::UnexpectedJniResultType(
+            "nativeHandle",
+            handle.type_name(),
+        )),
     }
 }
